+++ src/components/pages/CheckinCalendar.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect } from 'react'
 import { View } from 'react-native'
 import { Agenda, DateObject } from 'react-native-calendars'
 import { useDate } from '@/hooks/useDate'
@@ -6,7 +6,7 @@ import { useFoursquare } from '@/hooks/useFoursquare'
 import { useUtils } from '@/hooks/useUtils'
 import { COLORS } from '@/constants/Colors'
 import type { Checkin } from '@/types/Foursquare'
-import { Timeline } from '@/components/templates/Timeline.component'
+import { Timeline } from '@/components/templates/Timeline'
 import useColorScheme from '@/hooks/useColorScheme'
 import { logEvent } from '@/hooks/useAnalytics'
 import useAsyncFn from 'react-use/lib/useAsyncFn'
